requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
flask-mail
flask-login
pydantic
email-validator
psycopg2-binary
requests
cloudipsp
pyjwt
pytest
mock